Moved stack safety laws into SyncLaws
